+++ src/main/java/com/github/jlangch/venice/impl/types/VncAtom.java
@@ -23,6 +23,8 @@ package com.github.jlangch.venice.impl.types;
 
 import static com.github.jlangch.venice.impl.types.VncBoolean.False;
 
+import java.io.Closeable;
+import java.io.IOException;
 import java.util.concurrent.atomic.AtomicReference;
 
 import com.github.jlangch.venice.VncException;
@@ -33,7 +35,7 @@ import com.github.jlangch.venice.impl.util.MetaUtil;
 import com.github.jlangch.venice.impl.util.Watchable;
 
 
-public class VncAtom extends VncVal implements IDeref {
+public class VncAtom extends VncVal implements IDeref, Closeable {
 
     public VncAtom(final VncVal value) {
         super(Constants.Nil);
@@ -128,6 +130,25 @@ public class VncAtom extends VncVal implements IDeref {
 
     public void removeWatch(final VncKeyword name) {
         watchable.removeWatch(name);
+    }
+
+    @Override
+    public void close() throws IOException {
+        for(;;) {
+            final VncVal oldVal = deref();
+
+            if (oldVal instanceof Closeable) {
+               if (state.compareAndSet(oldVal, oldVal)) {
+                  try {
+                      ((Closeable)oldVal).close();
+                  }
+                  catch (Exception ex) {
+                     // best effort
+                  }
+                  return;
+              }
+            }
+        }
     }
 
     @Override